+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v0.7.0
 
+- Adds `Process.then` as an analogue of shell `P ; Q`.
 - Adds `CompletedProcess.ok` as a return-boolean version of `.check`.
 
 ## v0.6.1